Upload functions moved to Utils
